move get_fragments away from quality file
